chore: move FirebaseAISamples.kt to top-level dir
